pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "findevil"
version = "0.1.0"
description = "IABF Agent - Autonomous DFIR on SANS SIFT Workstation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "dhyabi2"},
]
dependencies = [
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "anyio>=4.0.0",
]

[project.scripts]
findevil = "main:main"

[tool.setuptools.packages.find]
include = ["agent*", "mcp_server*"]
